pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["dddops <Eopseum0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
ipython = "^9.6.0"
django-extensions = "^4.1"
django-cleanup = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
